add: 입찰 등록 자동 메세지

+++ backend/kickauction/src/main/java/com/sports/kickauction/dto/BizRegisterDTO.java
@@ -7,7 +7,7 @@ import lombok.Data;
 @Builder
 public class BizRegisterDTO {
     private int ono;          // 견적 요청 ID
-    private int price;         // 제안 가격
+    private Long price;         // 제안 가격
     private String bcontent;   // 상세 정보
     private String banswer;    // 요청 답변
 }
